CHANGES
I added the Time Conversion class, and moved all of the methods into
that class. I then added the user-input-oriented methods into the
TimeConversion class. Im currently working on the conversion method,
which will actually convert the units. But I need to find an efficient
method of doing so. So any advice is greatly appreciated.